Rewrite LinkedeventsImporter for Activity Streams
Basically a rewrite of the LinkedeventsImporter that iterates over all
the events that have been modified since some datetime (placeholder is
"today"). It outputs only the planned, relevant parts. Note that while
it now uses the include parameter to greatly increase speed compared to
its predecessor, it cannot do that with the publisher field yet due to
limitations of the Linkedevents API, however this should not be a
problem as the amount of events seems to be at most a few hundred per
day.

Closes #17